+++ cli/src/test/java/com/expediagroup/streamplatform/streamregistry/cli/command/ApplyTest.java
@@ -49,7 +49,7 @@ import com.expediagroup.streamplatform.streamregistry.state.model.specification.
 public class ApplyTest {
   private final CommandLine underTest = new CommandLine(new Apply());
 
-  private final String[] graphqlOptions = {"--streamRegistryUrl=streamRegistryUrl", "--username=username", "--password=password"};
+  private final String[] graphqlOptions = {"--streamRegistryUrl=streamRegistryUrl", "--streamRegistryUsername=streamRegistryUsername", "--streamRegistryPassword=streamRegistryPassword"};
   private final String[] configurationOptions = {"--description=description", "--tag=a:b", "--tag=c:d", "--type=type", "--configuration={\"e\":\"f\"}"};
 
   private final ObjectMapper mapper = new ObjectMapper();
@@ -93,8 +93,8 @@ public class ApplyTest {
   public void zone() {
     ParseResult result = underTest.parseArgs("zone",
         "--streamRegistryUrl=streamRegistryUrl",
-        "--username=username",
-        "--password=password",
+        "--streamRegistryUsername=streamRegistryUsername",
+        "--streamRegistryPassword=streamRegistryPassword",
         "--description=description", "--tag=a:b", "--tag=c:d", "--type=type", "--configuration={\"e\":\"f\"}",
         "--zone=zone");
     assertEvent(result, zoneKey);
@@ -163,7 +163,7 @@ public class ApplyTest {
 
   private void assertGraphQLOptions(GraphQLEventSenderAction action) {
     assertThat(action.getStreamRegistryUrl(), is("streamRegistryUrl"));
-    assertThat(action.getUsername(), is("username"));
-    assertThat(action.getPassword(), is("password"));
+    assertThat(action.getStreamRegistryUsername(), is("streamRegistryUsername"));
+    assertThat(action.getStreamRegistryPassword(), is("streamRegistryPassword"));
   }
 }